+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 * Prefer `calloc()` to `malloc()` in NDK code
   [#1320](https://github.com/bugsnag/bugsnag-android/pull/1320)
 
+* Ensure correct value always collected for activeScreen
+  [#1322](https://github.com/bugsnag/bugsnag-android/pull/1322)
+
 ## 5.10.0 (2021-07-14)
 
 ### Enhancements